+++ pages/_app.tsx
@@ -22,7 +22,7 @@ export default function App({ Component, pageProps }: AppProps) {
       <ServiceWorkerRegistration />
       <Component {...pageProps} />
       <Toaster
-        position="top-right"
+        position="top-center"
         toastOptions={{
           duration: 4000,
           style: {